Added damage reduction for E
Damage reduction is not settable via the menu. The default value of 20
should be okay, if you don't have any problems with 0 you can use that
aswell. Some might have issues that the monsters are still alive with
like 5-10 HP after using the E, so this will fix it for you.