Added AA+E on minions in combo
This was requested by "God" on the forums:

"Suggestion if it doesn't do it currently: When you have at least one
Rend stack on an enemy champion and on a minion, as long as the minion
will die, activate E. This way if done properly you can get more damage
out of E."

Basically this has been added to combo mode now. Might add it to harass
mode aswell.