Improved ult targetting and added single spell casts
Single spell casting can be used if you just wanna throw a E or W at
your enemy without any mode being active. This is not a "hold key"
feature, you must release the button in order for it to trigger, if it
does not shoot, it would not have it the enemy.